// src/app/model/OpeGuia.ts
export class OpeGuia {
    idOpeGuia: number;
    idOpeTipoGuia: number;
    idMaeDirecciones: number;
    idOpeTipoNacGuia: number;
    idOpeTipoJornadaGuia: number;
    nombre: string;
    rut: string;
    fechaNacimiento: string;
    telefonoFijo: string;
    telefonoMovil: string;
    email: string;
    profesion: string;
    cursos: string;
    primerosAuxilios: boolean;
    vigencia: string;
    regulares: boolean;
    privadoIndividual: boolean;
    privadoGrupal: boolean;
    tarifaTuristour: boolean;
    contactoEmergencia: string;
    contactoTelefonoEmergencia: string;
    centroAsistencia: string;
    enfermedades: string;
    alergias: string;
    observaciones: string;
    indicadorGuia: number;
    rankingGeneral: number;
    rankingIdiomas: number;
    rankingTerreno: number;
    estado: boolean;
    idMaeEmpresaGrupo: number;

}