// src/app/model/OpeServicio.ts
export class OpeServicio {
    idOpeServicio: number;
    idOpeVehiculo: number;
    idOpeConductor: number;
    idOpeGuia: number;
    idClieCliente_OpeServicio: number;
    idProvProveedores: number;
    idSegUsuario_OpeServicio: number;
    idOpeTipoServicio: number;
    idOpeTipoEstadoServicio: number;
    pasajero_OpeServicio: string;
    descripcion: string;
    montoTarifa: string;
    origen: string;
    destino: string;
    kmFin: number;
    noVueloHora: string;
    kmInicio: number;
    noPasajeros: number;
    noPasajerosPagados: number;
    noPasajerosReservados: number;
    observacion: string;
    fechaHoraInicio: string;
    fechaHoraFin: string;
    fechaHoraInicioReal: string;
    fechaHoraFinReal: string;
    idMaeEmpresaGrupo_OpeServicio: number;
    idOpeTipoVehiculo: number;


}