// src/app/model/Pasajero.ts
export class Pasajero {
    idOpeServicioDetalle: number;
    idOpeServicio: number;
    idMaeEmpresaGrupo: number;
    idOpeTipoServicio: number;
    idVentaTipoPago: number;
    idMaeIdiomas: number;
    chequeo: number;
    idOpeAcercamientoDetalle: number;
    idVentaRegularDetalle: number;
    pasajero: string;
    origen: string;
    destino: string;
    fechaHoraInicio: string;
    montoTarifa: string;
    noPaxAdulto: number;
    noPaxChild: number;
    noPasajerosPagados: number;
    noPasajerosReservados: number;
    direccionOrigen: string;
    direccionDestino: string;
    noHabitacion: string;
    comidas: number;
    entradas: number;
    tourVinos: number;
    otros: number;
    idVenEstadoWF: number;
    clienteIndicaPago: boolean;
    noVueloHora: string;
    observaciones: string;
    idProdSubTipoProducto: number;
    incluyeObs: boolean;
    estado: boolean;
    incluyeGuia: boolean;
    incluyeAgua: boolean;
    incluyeEntradas: boolean;
    incluyeTourVinos: boolean;
    incluyeComidas: boolean;
    incluyeOtros: boolean;
    idVentaRegular: number;
    idVentaRegularDetalleItinerario: number;
    idProdCosto: number;
    idProdProducto: number;
    idProdPrograma: number;
    clienteOperador: string;
    idMaeIdiomasServicio: number;
    contacto: string;
    esRecogida: boolean;
    fechaHoraRecogida: string;
    recogidaObs: string;

}